fix(cosh-ng): [core,shell] harden compaction flow
- keep pending and active cancellation revision-bound to prevent retriggers
- move process lifecycle coverage to raw CLI and enforce test layout
- isolate compaction runtime, preflight, and engine tests by responsibility
